More

+++ pretrain/web/refined_web.py
@@ -1,21 +1,23 @@
 
 import math
+import multiprocess
 import os
 import regex as re
-import numpy as np
+from collections import Counter
 
+import argparse
 import boto3
 import botocore
+import numpy as np
 from datasets import concatenate_datasets, load_dataset
-
 
 MIN_DOC_TOKENS = 50
 
 # Use RefinedWeb
 WEB_DIR = '/weka/home-griffin/clinical_pile/refined_web'
-os.makedirs(WEB_DIR, exist_ok=True)
+SHARD_DIR = os.path.join(WEB_DIR, 'shards')
+os.makedirs(SHARD_DIR, exist_ok=True)
 BUCKET_NAME = 'pile-everything-west' # replace with your bucket name
-TARGET_PAGES = 10000000
 
 
 def download(s3, key):
@@ -41,6 +43,17 @@ def process(row):
 
 
 if __name__ == '__main__':
+    parser = argparse.ArgumentParser('Refined Web Download from S3 and sample for target token count.')
+
+    parser.add_argument('--target_num_pages', default=10000000, type=int, help='Total number of webpages to sample.')
+
+    parser.add_argument('--chunk', default=None, type=int)
+    parser.add_argument('--num_chunks', default=10, type=int)
+
+    args = parser.parse_args()
+
+    from datasets import load_from_disk
+
     s3 = boto3.resource('s3')
     key = f'falcon-refinedweb/'
 
@@ -54,47 +67,79 @@ if __name__ == '__main__':
             keys.append(obj.key)
 
     num_shards = len(keys)
-    pages_per_shard = math.ceil(TARGET_PAGES / num_shards)
+    pages_per_shard = math.ceil(args.target_num_pages / num_shards)
     print(f'Getting {pages_per_shard} pages per shard (# {num_shards})')
 
-    datasets = []
+    if args.chunk is not None:
+        keys = np.array_split(keys, args.num_chunks)[args.chunk]
+
     for key in keys:
         shard = int(re.search(r'\d+', key).group())
+
         print(f'Starting with shard {shard}...')
-        # Use boto3 to download the dataset
-        shard_local_fn = download(s3, key)
+        shard_dir = os.path.join(SHARD_DIR, f'{shard}_hf')
 
-        # Use HF to load it
-        dataset = load_dataset('parquet', data_files=shard_local_fn, split='train')
+        if os.path.exists(shard_dir):
+            print(f'Found existing processed shard at {shard_dir}. Skipping...')
+        else:
+            # Use boto3 to download the dataset
+            shard_local_fn = download(s3, key)
 
-        idxs = np.arange(len(dataset))
-        np.random.shuffle(idxs)
+            # Use HF to load it
+            dataset = load_dataset('parquet', data_files=shard_local_fn, split='train')
 
-        sample = dataset.select(idxs[:pages_per_shard])
+            idxs = np.arange(len(dataset))
+            np.random.shuffle(idxs)
 
-        # Computing Number of Tokens / Document
-        sample = sample.map(
-            lambda row: {'num_tokens': len(re.split(r'\W+', row['content']))},
-            num_proc=64
-        )
+            sample = dataset.select(idxs[:pages_per_shard])
 
-        sample = sample.filter(lambda row: row['num_tokens'] >= MIN_DOC_TOKENS)
+            # Computing Number of Tokens / Document
+            sample = sample.map(
+                lambda row: {'num_tokens': len(re.split(r'\W+', row['content']))},
+                num_proc=multiprocess.cpu_count()
+            )
 
-        sample = sample.map(
-            lambda row: {'shard': shard, 'text': process(row), 'id': row['url']},
-            remove_columns=['image_urls', 'content']
-        )
+            sample = sample.filter(
+                lambda row: row['num_tokens'] >= MIN_DOC_TOKENS,
+                num_proc=multiprocess.cpu_count()
+            )
 
-        datasets.append(sample)
+            sample = sample.map(
+                lambda row: {'shard': shard, 'text': process(row), 'id': row['url']},
+                remove_columns=['image_urls', 'content'],
+                num_proc=multiprocess.cpu_count()
+            )
 
-        # Remove shard
-        print(f'Done with {shard_local_fn}. Removing it...')
-        os.remove(shard_local_fn)
+            print(f'Saving {len(sample)} pages to {shard_dir}')
+            sample.save_to_disk(shard_dir)
+
+            # Avoid out of disk errors
+            print('Cleaned up ' + str(sample.cleanup_cache_files()) + ' cache files')
+
+            # Remove shard
+            print(f'Done with {shard_local_fn}. Removing it...')
+            os.remove(shard_local_fn)
+
+    datasets = []
+    for subdir in os.listdir(SHARD_DIR):
+        full_path = os.path.join(SHARD_DIR, subdir)
+        print(f'Adding {full_path} to full dataset')
+        datasets.append(load_from_disk(full_path))
 
     datasets = concatenate_datasets(datasets)
+
+    dup_ids = set([id[0] for id in Counter(datasets['id']).most_common() if id[1] > 1])
+
+    print(f'Found {len(dup_ids)} duplicate URLs. Removing all instances')
+    prev_n = len(datasets)
+    datasets = datasets.filter(
+        lambda row: row['id'] not in dup_ids,
+        num_proc=multiprocess.cpu_count()
+    )
+    new_n = len(datasets)
+    print(f'Removed {prev_n - new_n} web pages with duplicate URLs.')
 
     out_dir = os.path.join(WEB_DIR, 'dataset_hf')
     total_toks = sum(datasets['num_tokens'])
     print(f'Saving {len(datasets)} web pages with {total_toks} total tokens to {out_dir}')
     dataset.save_to_disk(out_dir)
-
